+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### 20141002
+* Requirements files
+
 ### 20141001
 * Added deb installers to Text Editors